move delete button to item box
