+++ web/src/components/item-box.js
@@ -9,7 +9,8 @@ import Fab from '@mui/material/Fab';
 import NavigationArrowForwardIcon from '@mui/icons-material/ArrowForward';
 import NavigationArrowBackIcon from '@mui/icons-material/ArrowBack';
 import ListIcon from '@mui/icons-material/List';
-import EditorModeEditIcon from '@mui/icons-material/Edit';
+import EditIcon from '@mui/icons-material/Edit';
+import DeleteIcon from '@mui/icons-material/Delete';
 import Typography from '@mui/material/Typography';
 import ElevationBox from './elevation-box';
 import PanoramaBox from './panorama-box';
@@ -20,7 +21,9 @@ import Chip from '@mui/material/Chip';
 import Avatar from '@mui/material/Avatar';
 import NoSsr from '@mui/material/NoSsr';
 import SwipeableViews from 'react-swipeable-views';
+import fetchWithAuth from '../fetch_with_auth';
 import { idToUrl } from '../app';
+import { push } from '@lagunovsky/redux-react-router';
 
 const ItemBox = () => {
     const [tabValue, setTabValue] = useState(0);
@@ -41,6 +44,19 @@ const ItemBox = () => {
     });
     const indexCHangeCB = useCallback(index => {
         setTabValue(index);
+    });
+    const handleDelete = useCallback(async e => {
+        if (confirm('Are you sure to delete?')) {
+            try {
+                const response = await fetchWithAuth('/api/destroy/' + selectedItem.id);
+                if (!response.ok) {
+                    throw Error(response.statusText);
+                }
+                dispatch(push('/?forceFetch=1'));
+            } catch (error) {
+                alert(error);
+            }
+        }
     });
 
     const data = selectedItem;
@@ -81,7 +97,10 @@ const ItemBox = () => {
                 <IconButton disabled={!nextUrl} component={Link} to={nextUrl || ''} size="large"><NavigationArrowBackIcon /></IconButton>
                 <IconButton disabled={!prevUrl} component={Link} to={prevUrl || ''} size="large"><NavigationArrowForwardIcon /></IconButton>
                 {
-                    data && currentUser && data.uid && currentUser.uid == data.uid ? (<IconButton onClick={handleEdit} size="large"><EditorModeEditIcon /></IconButton>) : null
+                    data && currentUser && data.uid && currentUser.uid == data.uid ? (<IconButton onClick={handleEdit} size="large"><EditIcon /></IconButton>) : null
+                }
+                {
+                    data && currentUser && data.uid && currentUser.uid == data.uid ? (<IconButton onClick={handleDelete} size="large"><DeleteIcon /></IconButton>) : null
                 }
                 <Typography variant="h6" sx={{ fontSize: '100%' }}>{ title || 'not found'}</Typography>
                 <Box sx={{ textAlign: 'right'}}>{
